+++ src/static-files-generator/static-file-generator.controller.ts
@@ -1,7 +1,7 @@
 import { Controller, Get } from '@nestjs/common';
 import { StaticFileGeneratorService } from './static-file-generator.service';
 
-@Controller('static-file')
+@Controller('static-files')
 export class StaticFileGeneratorController {
   constructor(private readonly staticFileGeneratorService: StaticFileGeneratorService) {}